requirements.txt
openai==0.28.0
pillow==10.4.0
tqdm==4.66.4
cython 
scipy
opencv-python
scikit-image
timm
einops
pandas
imgaug
h5py
av
re
ast
time
json
math
typing
logging
warnings
base64
easydict
librosa
threading
matplotlib
itertools
soundfile
multiprocessing
